requirements.txt
gkeepapi==0.11.5
python-dotenv==0.10.1
